pyproject.toml
[project]
name = "kted-edu-1day"
version = "0.1.0"
description = "1일차"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-cli>=0.0.36",
    "langchain-community[all]>=0.3.26",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langserve[all]>=0.3.1",
    "pandas>=2.3.0",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
]

[tool.uv.workspace]
members = [
    "ktds-edu",
]
